Move Junit launcher to its own module as external projects will depend on this module and commons module classes should not be visible to them
