+++ samples/streaming/src/test/java/org/eclipse/dataspaceconnector/transfer/demo/protocols/DemoProtocolsTransferExtensionTest.java
@@ -14,7 +14,7 @@
 
 package org.eclipse.dataspaceconnector.transfer.demo.protocols;
 
-import org.eclipse.dataspaceconnector.common.testfixtures.EdcExtension;
+import org.eclipse.dataspaceconnector.junit.launcher.EdcExtension;
 import org.eclipse.dataspaceconnector.spi.monitor.Monitor;
 import org.eclipse.dataspaceconnector.spi.security.Vault;
 import org.eclipse.dataspaceconnector.spi.security.VaultResponse;
